refactor: reorganize signer modules and remove legacy dependencies
Privy and Turnkey as separate crates wasn't needed, we would never publish the crates themselves and they were tightly coupled with kora_lib

- Removed `kora-privy` and `kora-turnkey` crates, consolidating their functionality into the `signer` module.
- Introduced new `privy` and `turnkey` submodules within the `signer` module for better organization.
- Added utility functions for hex and byte conversions in the `signer` module.
