+++ Cargo.toml
@@ -1,12 +1,5 @@
 [workspace]
-members = [
-    "crates/lib",
-    "crates/rpc",
-    "crates/cli",
-    "crates/turnkey",
-    "crates/privy",
-    "tests",
-]
+members = ["crates/lib", "crates/rpc", "crates/cli", "tests"]
 resolver = "2"
 
 [workspace.lints.rust]
@@ -75,15 +68,15 @@ jup-ag = "0.8.0"
 toml = "0.8.19"
 spl-token = { version = "7.0.0", features = ["no-entrypoint"] }
 spl-token-2022 = { version = "8.0.0", features = ["no-entrypoint"] }
-spl-associated-token-account = { version = "6.0.0", features = ["no-entrypoint"] }
+spl-associated-token-account = { version = "6.0.0", features = [
+    "no-entrypoint",
+] }
 chrono = "0.4.39"
 hex = "0.4.3"
 p256 = "0.13.3"
 redis = { version = "0.24.0", features = ["tokio-comp", "connection-manager"] }
 deadpool-redis = "0.14.0"
 vaultrs = "0.7.3"
-kora-turnkey = { path = "crates/turnkey", version = "1.0.2" }
-kora-privy = { path = "crates/privy", version = "1.0.2" }
 utoipa = { version = "4.2.0", features = ["yaml", "chrono"] }
 hmac = "0.12.1"
 sha2 = "0.10.9"
