+++ crates/lib/src/signer/mod.rs
@@ -1,7 +1,10 @@
 pub mod init;
 pub mod keypair_util;
+pub mod privy;
 pub mod signer;
 pub mod solana_signer;
+pub mod turnkey;
+pub mod utils;
 pub mod vault_signer;
 
 pub use keypair_util::KeypairUtil;
