+++ crates/lib/src/signer/privy/mod.rs
@@ -1,0 +1,3 @@
+pub mod config;
+pub mod signer;
+pub mod types;
