+++ crates/lib/src/signer/utils.rs
