--- crates/privy/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "kora-privy"
-version = "1.0.3"
-edition = "2021"
-description = "A Rust library for Privy signing"
-license = "MIT"
-
-[dependencies]
-solana-sdk = "2.1"
-reqwest = { version = "0.11", features = ["json"] }
-base64 = "0.21"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-tokio = { version = "1", features = ["full"] }
-thiserror = "1.0"
-async-trait = "0.1"
-anyhow = "1.0"
-bincode = "1.3"
-
-[dev-dependencies]
-clap = { version = "4.0", features = ["derive", "env"] }
-dotenv = "0.15"
-env_logger = "0.10"
-
-[lib]
-name = "kora_privy"
-path = "src/mod.rs"
